refact: extract resources helpers to utils
