+++ src/utils/index.ts
@@ -1,4 +1,7 @@
-import { IFileOrFolder } from "../types"
+import chain from "ramda/es/chain"
+import partition from "ramda/es/partition"
+
+import { IFile, IFileOrFolder, IFolder } from "../types"
 
 export function traversePaths(
   [head, ...tail]: string[],
@@ -46,4 +49,61 @@ export function traversePathsToFiles(
     path,
     children: [traversePathsToFiles(tail, path, 1)],
   }
+}
+
+function generateFolder({
+  title,
+  path,
+  targets,
+}: {
+  title: IFolder["title"]
+  path: IFile["path"]
+  targets: IFolder[]
+}): IFolder {
+  const children = join(chain(target => target.children, targets))
+
+  return {
+    title,
+    type: "folder",
+    path,
+    children: children.sort((a, b) =>
+      a.title.split("/").length > b.title.split("/").length
+        ? -1
+        : a.title.localeCompare(b.title)
+    ),
+  }
+}
+
+function generateFile({
+  title,
+  path,
+}: {
+  title: IFile["title"]
+  path: IFile["path"]
+}): IFile {
+  return {
+    title,
+    type: "file",
+    path,
+  }
+}
+
+export function join([head, ...tail]: IFileOrFolder[]): IFileOrFolder[] {
+  if (!head) {
+    return []
+  }
+
+  const [similarFs, remaining] = partition(
+    obj => obj.title === head.title && obj.type === head.type,
+    tail
+  )
+  const targets = [head, ...similarFs]
+  const { title, path } = head
+
+  const current =
+    head.type === "folder"
+      ? generateFolder({ title, path, targets: targets as IFolder[] })
+      : generateFile({ title, path })
+
+  return [current, ...join(remaining)]
 }
